experimental wasm support

+++ gamedata/shaders/city_mesh.glsl
@@ -10,8 +10,8 @@ uniform usampler2D tex_2; // palette indices table
 // attributes
 in vec3 in_pos0;
 in vec3 in_texcoord0;
-in int in_color0; // remap index
-in int in_color1; // transparency flag
+in uint in_color0; // remap index
+in uint in_color1; // transparency flag
 
 // pass to fragment shader
 out vec3 Texcoord;
@@ -25,18 +25,18 @@ const float MeshHeightModifier = -0.15; // shift the geometry level slightly bel
 void main() 
 {
 	Texcoord = in_texcoord0;
-    Transparency = in_color1;
+    Transparency = float(in_color1);
 
     // get real block tile index
-    BlockTextureIndex = texelFetch(tex_1, ivec2(int(Texcoord.z + 0.5), 0), 0).r;
+    BlockTextureIndex = float(texelFetch(tex_1, ivec2(int(Texcoord.z + 0.5), 0), 0).r);
 
     // get palette index for block tile
-    PaletteIndex = texelFetch(tex_2, ivec2(int(4.0 * BlockTextureIndex + in_color0), 0), 0).r;
+    PaletteIndex = float(texelFetch(tex_2, ivec2(int(4.0 * BlockTextureIndex + float(in_color0)), 0), 0).r);
 
     vec4 vertexPosition = view_projection_matrix * vec4(
 		in_pos0.x, 
 		in_pos0.y + MeshHeightModifier, 
-		in_pos0.z, 1.0f);
+		in_pos0.z, 1.0);
 
     gl_Position = vertexPosition;
 }
@@ -62,15 +62,15 @@ out vec4 FinalColor;
 void main()
 {
     // get color index in palette
-    float pal_color = texture(tex_0, vec3(Texcoord.x, Texcoord.y, BlockTextureIndex)).r;
+    float pal_color = float(texture(tex_0, vec3(Texcoord.x, Texcoord.y, BlockTextureIndex)).r);
 
     if (Transparency > 0.5 && pal_color < 0.5) // old school alpha test
         discard;
 
     // fetch pixel color at least
     vec4 pixelColor = texelFetch(tex_3, ivec2(int(pal_color), int(PaletteIndex)), 0);
-    pixelColor.a = 1.0f;
-    FinalColor = clamp(pixelColor, 0.0f, 1.0f);
+    pixelColor.a = 1.0;
+    FinalColor = clamp(pixelColor, 0.0, 1.0);
 }
 
 #endif
